env: adjusting docker-compose and adding scripts

+++ database/create-database.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS vote (
   id INTEGER AUTO_INCREMENT NOT NULL PRIMARY KEY,
   agenda_id INTEGER NOT NULL,
   vote VARCHAR(3) NOT NULL,
-  associated_id INTEGER NOT NULL,
+  associated_id BIGINT NOT NULL,
   associated_cpf VARCHAR(14),
   FOREIGN KEY (agenda_id) REFERENCES agenda(id),
   UNIQUE (agenda_id, associated_id)
